feat: huge rework of datamodels

+++ crates/aurorite-server/src/utils/formulas.rs
@@ -1,11 +1,11 @@
-pub fn get_modification(value: i16) -> i16 {
+pub fn get_modification(value: u8) -> i8 {
     if value <= 11 {
-        (value - 11) / 2
+        (value as i8 - 11) / 2
     } else {
-        (value - 10) / 2
+        (value as i8 - 10) / 2
     }
 }
 
-pub fn get_proficiency_bonus(level: i16) -> i16 {
+pub fn get_proficiency_bonus(level: u8) -> u8 {
     2 + ((level - 1) / 4)
 }
